pebblo_saferetriever/langchain/identity-rag-sharepoint/requirements.txt
python-dotenv
requests
langchain
o365
pymupdf
langchain_openai